+++ tests/test_DataSet_class.py
@@ -403,15 +403,16 @@ def test_update_manifest_does_nothing_if_not_persisted():
 def test_decriptive_metadata_property(tmp_dir_fixture):  # NOQA
     from dtoolcore import DataSet
     dataset = DataSet('my_dataset')
-    assert dataset.descriptive_metadata == {}
+    assert dataset.raw_descriptive_metadata == ""
 
     dataset.persist_to_path(tmp_dir_fixture)
-    assert dataset.descriptive_metadata == {}
+    assert dataset.raw_descriptive_metadata == ""
 
     with open(dataset.abs_readme_path, "w") as fh:
         fh.write("---\n")
         fh.write("project: my_project\n")
-    assert dataset.descriptive_metadata == {"project": "my_project"}
+    assert dataset.raw_descriptive_metadata ==  \
+        open(dataset.abs_readme_path, "r").read()
 
 
 def test_identifiers_property():